notes fix
